requirements.txt
Pillow==9.1.1
streamlit==1.9.0
streamlit_option_menu==0.3.2
